requirements.txt
torch
diffusers>=0.14.0
transformers
controlnet-aux
rembg
Pillow
numpy
opencv-python
